moving batch related sources into a read package
