+++ ontology-rest/src/main/java/uk/ac/ebi/quickgo/ontology/traversal/read/OntologyTraversalConfig.java
@@ -1,8 +1,9 @@
-package uk.ac.ebi.quickgo.ontology.traversal;
+package uk.ac.ebi.quickgo.ontology.traversal.read;
 
 import uk.ac.ebi.quickgo.common.batch.listener.LogJobListener;
 import uk.ac.ebi.quickgo.common.batch.listener.LogStepListener;
 import uk.ac.ebi.quickgo.common.batch.listener.SkipLoggerListener;
+import uk.ac.ebi.quickgo.ontology.traversal.OntologyGraph;
 
 import java.io.IOException;
 import java.util.ArrayList;
